requirements.txt
flask
flask_sqlalchemy
flask_wtf
python-dotenv
email_validator 